requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
psycopg2-binary==2.9.11
numpy==2.3.4
scipy==1.16.2
pydantic==2.12.2
typing-extensions==4.15.0
requests==2.32.5
